updating proj

+++ src/main/java/tikal/analyzer/handlers/mongo/MongoHandlerImpl.java
@@ -74,7 +74,7 @@ public class MongoHandlerImpl implements MongoHandler {
 		MongoSlackRawData result = new MongoSlackRawData();
 		result.setUserName(data.getString("user"));
 		result.setChannel(data.getString("channel"));
-		result.setTimestamp(data.getLong("timestamp"));
+		result.setTimestamp(data.getDouble("timestamp").longValue());
 
 		return result;
 	}
